% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volpak_tree.R, R/volpak_tree_methods.R
\name{volpak_tree}
\alias{volpak_tree}
\alias{volpak_vol_to_tdub}
\alias{volpak_vol_to_hag}
\alias{volpak_total_vol}
\alias{volpak_get_hag}
\alias{volpak_get_diam}
\alias{volpak_tree_height}
\alias{volpak_stump_radius}
\alias{volpak_stump_vol}
\alias{volpak_ground_radius}
\alias{volpak_print}
\alias{volpak_vol_above_top}
\title{R Interface to Volpak routines}
\usage{
volpak_tree(heights, diameters, total.height, stump.height = 0.15)

volpak_vol_to_tdub(tree, tdubs, abovestump = FALSE)

volpak_vol_to_hag(tree, hags, abovestump = FALSE)

volpak_total_vol(tree, abovestump = FALSE)

volpak_get_hag(tree, search.diam)

volpak_get_diam(tree, search.hag)

volpak_tree_height(tree)

volpak_stump_radius(tree)

volpak_stump_vol(tree)

volpak_ground_radius(tree)

volpak_print(tree)

volpak_vol_above_top(tree)
}
\arguments{
\item{heights}{The height above ground corresponding to each diameter
measurement.}

\item{diameters}{The under-bark diameter measured along the tree.}

\item{total.height}{The total height of the tree.}

\item{stump.height}{The height (in metres above ground) of the stump.}

\item{abovestump}{Should calculated volumes exclude the 15cm stump.}
}
\value{
\code{volpak_tree} returns NULL, \code{volpak_total_vol} returns a
single numeric value, all remaining functions return a numeric vector with
the same length as the input vectors.
}
\description{
\code{volpak_tree} returns an object of type \linkS4class{volpak_tree}
containing the information stored in the \code{Tree} class, used by volpak.
\code{volpak_total_vol} calculates the total under-bark volume between a 15cm
stump and the top of the tree. \code{volpak_get_hag} calculates the height
above ground at which the tree has the requested diameter.
\code{volpak_vol_to_tdub} calculates the volume between a 15cm stump and the
under-bark diameter provided. \code{volpak_vol_to_hag} calculates the volume
between a 15cm stump and the height above ground provided.
}
\details{
The volpak library is used by GAME to produce volume estimates, as well as
interpolated heights and diameters, for trees that have been measured as park
of HQPlantations DST program.

The interface between R and volpak has been written for simplicity, as such
all inputs must be relevant to a single tree. An interface to the volpak
initialisation routine has been provided, however each of the exported
functions in the package will also initialise volpack internally.
}
\examples{
ht <- c(0.5, 1.3, 5, 8, 11, 14, 17, 20)
diam <- round(10/sqrt(ht), 1)
tht <- 22

volpak_tree(ht, diam, tht)
volpak_total_vol(ht, diam, tht)
volpak_get_hag(c(5, 2.8), ht, diam, tht)
volpak_vol_to_tdub(c(5, 2.8), ht, diam, tht)
volpak_vol_to_hag(c(10, 13), ht, diam, tht)
}
